LocalBlobAccess: Factor out logic for determining to create new blocks
OldCurrentNewLocationBlobMap currently has this feature that it can
use an elevated number of "new" blocks during startup to spread out data
aggressively.

This feature is great for the CAS, but it turns out to be problematic
for the AC and ICAS. It makes it hard to reliably overwrite AC and ICAS
entries, as they are not guaranteed to be placed in a block that is at
least as new as the one holding the existing object.